Changed GID input / create ui
